Thin idm-keyrock image by using shallow git clones.
We don't need the full repositories history on the image, so by using
shallow clones of the git repositories we can reduce the size of the
final image.  If we need the full history at a later stage, we can pull
it and convert the clone to a full one.